Adds the ability to upload local files and associate them with an attachment field.
The new upload_file method allows a local file to be uploaded to NocoDB and assigned
to a field with the attachment type. The upload is done in two steps: First, the file
is transferred to the NocoDB server using the upload endpoint
(https://apis.nocodb.com/#tag/Storage/operation/storage-upload) of the Meta-Api. Then,
with the help of table_row_update, the specified field is linked to the uploaded file.